# CAPT-1331 GIAS - 'Select your school' not listing school searched for with post code only